Split task/mod.rs into multiple files.
